pyproject.toml
[tool.poetry]
name = "loop-assign"
version = "0.1.0"
description = ""
authors = ["Sai Suvanth <saisuvanth@gmail.com>"]
readme = "README.md"
packages = [{include = "loop_assign"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = ">=1.4.42,<1.5"
databases = "^0.7.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
psycopg2 = "^2.9.6"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

